Refactor in SyntaxDefinition and related classes. Minor improvements in CPP, CSS, xit syntax highlighting. AI Assistant / LLM Chat now can restore previous opened chats and states (last model used, panel split), can have private conversations, can clone a previous conversation, plus some visual improvements. Fixed a bug in tab widget splitting.

+++ include/eepp/ui/doc/syntaxtokenizer.hpp
@@ -55,8 +55,8 @@ struct SyntaxState {
 	Uint8 state[MAX_SUB_SYNTAXS]{ SYNTAX_TOKENIZER_STATE_NONE, SYNTAX_TOKENIZER_STATE_NONE,
 								  SYNTAX_TOKENIZER_STATE_NONE, SYNTAX_TOKENIZER_STATE_NONE };
 
-	// 8 bits per language (language index) - max 4 sub-languages - max 255 languages
-	Uint8 langStack[MAX_SUB_SYNTAXS]{ 0, 0, 0, 0 };
+	// 16 bits per language (language index) - max 4 sub-languages - max 65k languages
+	Uint16 langStack[MAX_SUB_SYNTAXS]{ 0, 0, 0, 0 };
 
 	bool operator==( const SyntaxState& other ) {
 		return memcmp( this, &other, sizeof( SyntaxState ) ) == 0;
